Flask-API/requirements.txt
flask
flask-cors
jinja2
google-generativeai
python-dotenv
requests